+++ build-tools/xaprepare/xaprepare/Steps/Step_PrepareBundle.cs
@@ -84,7 +84,7 @@ namespace Xamarin.Android.Prepare
 					return false;
 				}
 
-				Log.DebugLine ("Moving unpacked bundle from {tempDir} to {Configurables.Paths.Bundle_InstallDir}");
+				Log.DebugLine ($"Moving unpacked bundle from {tempDir} to {Configurables.Paths.BundleInstallDir}");
 				Utilities.MoveDirectoryContentsRecursively (tempDir, Configurables.Paths.BundleInstallDir, resetFileTimestamp: true);
 			} finally {
 				Utilities.DeleteDirectorySilent (tempDir);
