+++ build-tools/xaprepare/xaprepare/xaprepare.csproj
@@ -1,4 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
@@ -249,6 +249,7 @@
     <Compile Include="Application\DetermineWindowsVersion.Windows.cs" />
     <Compile Include="Application\LlvmRuntime.Windows.cs" />
     <Compile Include="Application\Log.Windows.cs" />
+    <Compile Include="Application\MonoCrossRuntime.Windows.cs" />
     <Compile Include="Application\Utilities.Windows.cs" />
     <Compile Include="ConfigAndData\Configurables.Windows.cs" />
     <Compile Include="ConfigAndData\Dependencies\AndroidToolchain.Windows.cs" />
